refactor(go/pkg/prometheus): reorganize files by responsibility
Rename client.go -> scrape.go (the scraper) now that the transport lives in fetch.go, and client_test.go -> scrape_test.go to match. Rename model.go -> sample.go (the streamed Sample model), consistent with metric_family.go / metric_series.go. Split parse.go into the parse stage (parse.go: driver + orchestrator) and the assemble stage (assemble.go: typed-family folding), with the free label helpers in labels.go. Pure file reorganization; no logic or public-API change.
